+++ src/pages/index/index.js
@@ -1,2 +1,2 @@
-import Home from './Home'
+import Home from './home'
 export default Home
