+++ src/pages/legal/terms.js
@@ -1,5 +1,4 @@
 import React from 'react'
-
 function Terms(props) {
   return (
     <div className="terms page">
