+++ src/pages/shop/shop.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import Link from 'next/link'
-
 function Shop(props) {
   return (
     <div className="page shop">
